requirements.txt
zeep==4.2.1
boto3==1.28.2
python-daemon==3.0.1